requirements.txt
psycopg2-binary>=2.9.1
pandas>=1.3.0
pytest>=6.2
ipython>=7.31.0
requests>=2.26.0
dbt-core==1.7.4
dbt-postgres==1.7.4
geopy>=2.3.0
pycountry>=22.3.5
pycountry-convert>=0.7.2